LPS-54167 Move display classes and update
