+++ modules/apps/marketplace/marketplace-web/docroot/WEB-INF/liferay-portlet.xml
@@ -7,7 +7,7 @@
 		<icon>/store/icon.png</icon>
 		<control-panel-entry-category>marketplace</control-panel-entry-category>
 		<control-panel-entry-weight>2.0</control-panel-entry-weight>
-		<control-panel-entry-class>com.liferay.marketplace.store.portlet.StoreControlPanelEntry</control-panel-entry-class>
+		<control-panel-entry-class>com.liferay.marketplace.web.store.portlet.StoreControlPanelEntry</control-panel-entry-class>
 		<use-default-template>false</use-default-template>
 		<private-request-attributes>false</private-request-attributes>
 		<private-session-attributes>false</private-session-attributes>
@@ -23,7 +23,7 @@
 		<icon>/my_marketplace/icon.png</icon>
 		<control-panel-entry-category>marketplace</control-panel-entry-category>
 		<control-panel-entry-weight>3.0</control-panel-entry-weight>
-		<control-panel-entry-class>com.liferay.marketplace.store.portlet.StoreControlPanelEntry</control-panel-entry-class>
+		<control-panel-entry-class>com.liferay.marketplace.web.store.portlet.StoreControlPanelEntry</control-panel-entry-class>
 		<preferences-owned-by-group>false</preferences-owned-by-group>
 		<use-default-template>false</use-default-template>
 		<private-request-attributes>false</private-request-attributes>
@@ -40,7 +40,7 @@
 		<icon>/app_manager/icon.png</icon>
 		<control-panel-entry-category>marketplace</control-panel-entry-category>
 		<control-panel-entry-weight>1.0</control-panel-entry-weight>
-		<control-panel-entry-class>com.liferay.marketplace.appmanager.portlet.AppManagerControlPanelEntry</control-panel-entry-class>
+		<control-panel-entry-class>com.liferay.marketplace.web.appmanager.portlet.AppManagerControlPanelEntry</control-panel-entry-class>
 		<preferences-owned-by-group>false</preferences-owned-by-group>
 		<use-default-template>false</use-default-template>
 		<private-request-attributes>false</private-request-attributes>
