+++ modules/apps/marketplace/marketplace-web/docroot/WEB-INF/portlet.xml
@@ -4,7 +4,7 @@
 	<portlet>
 		<portlet-name>1</portlet-name>
 		<display-name>Store</display-name>
-		<portlet-class>com.liferay.marketplace.store.portlet.StorePortlet</portlet-class>
+		<portlet-class>com.liferay.marketplace.web.store.portlet.StorePortlet</portlet-class>
 		<init-param>
 			<name>view-template</name>
 			<value>/store/view.jsp</value>
@@ -35,7 +35,7 @@
 	<portlet>
 		<portlet-name>2</portlet-name>
 		<display-name>Purchased</display-name>
-		<portlet-class>com.liferay.marketplace.store.portlet.StorePortlet</portlet-class>
+		<portlet-class>com.liferay.marketplace.web.store.portlet.StorePortlet</portlet-class>
 		<init-param>
 			<name>view-template</name>
 			<value>/store/view.jsp</value>
@@ -66,7 +66,7 @@
 	<portlet>
 		<portlet-name>3</portlet-name>
 		<display-name>App Manager</display-name>
-		<portlet-class>com.liferay.marketplace.appmanager.portlet.AppManagerPortlet</portlet-class>
+		<portlet-class>com.liferay.marketplace.web.appmanager.portlet.AppManagerPortlet</portlet-class>
 		<init-param>
 			<name>view-template</name>
 			<value>/app_manager/view.jsp</value>
