--- modules/apps/marketplace/marketplace-web/docroot/WEB-INF/src/com/liferay/marketplace/store/portlet/StoreControlPanelEntry.java
@@ -1,40 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.marketplace.store.portlet;
-
-import com.liferay.marketplace.util.PortletPropsValues;
-import com.liferay.portal.model.Group;
-import com.liferay.portal.model.Portlet;
-import com.liferay.portal.security.permission.PermissionChecker;
-import com.liferay.portlet.BaseControlPanelEntry;
-
-/**
- * @author Ryan Park
- */
-public class StoreControlPanelEntry extends BaseControlPanelEntry {
-
-	@Override
-	public boolean hasAccessPermission(
-			PermissionChecker permissionChecker, Group group, Portlet portlet)
-		throws Exception {
-
-		if (!PortletPropsValues.MARKETPLACE_STORE_ENABLED) {
-			return false;
-		}
-
-		return permissionChecker.isOmniadmin();
-	}
-
-}
