+++ modules/apps/marketplace/marketplace-web/docroot/WEB-INF/src/com/liferay/marketplace/web/util/MarketplaceConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.marketplace.util;
+package com.liferay.marketplace.web.util;
 
 import com.liferay.portal.kernel.util.ReleaseInfo;
 import com.liferay.portal.kernel.util.StringPool;
