+++ modules/apps/marketplace/marketplace-web/docroot/WEB-INF/src/com/liferay/marketplace/web/util/MarketplaceLicenseUtil.java
@@ -12,18 +12,16 @@
  * details.
  */
 
-package com.liferay.marketplace.util;
+package com.liferay.marketplace.web.util;
 
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.util.FileUtil;
-import com.liferay.portal.kernel.util.MethodKey;
-import com.liferay.portal.kernel.util.PortalClassInvoker;
-import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.license.util.LicenseManagerUtil;
+import com.liferay.portal.license.util.LicenseUtil;
 
 import java.io.File;
 
@@ -48,26 +46,7 @@ public class MarketplaceLicenseUtil {
 		jsonObject.put("productEntryName", productEntryName);
 		jsonObject.put("serverId", Arrays.toString(getServerIdBytes()));
 
-		Thread currentThread = Thread.currentThread();
-
-		ClassLoader classLoader = currentThread.getContextClassLoader();
-
-		String response = null;
-
-		try {
-			currentThread.setContextClassLoader(
-				PortalClassLoaderUtil.getClassLoader());
-
-			MethodKey methodKey = new MethodKey(
-				"com.liferay.portal.license.util.LicenseUtil", "sendRequest",
-				String.class);
-
-			response = (String)PortalClassInvoker.invoke(
-				methodKey, jsonObject.toString());
-		}
-		finally {
-			currentThread.setContextClassLoader(classLoader);
-		}
+		String response = LicenseUtil.sendRequest(jsonObject.toString());
 
 		JSONObject responseJSONObject = JSONFactoryUtil.createJSONObject(
 			response);
@@ -88,24 +67,7 @@ public class MarketplaceLicenseUtil {
 	public static void registerOrder(String orderUuid, String productEntryName)
 		throws Exception {
 
-		Thread currentThread = Thread.currentThread();
-
-		ClassLoader classLoader = currentThread.getContextClassLoader();
-
-		try {
-			currentThread.setContextClassLoader(
-				PortalClassLoaderUtil.getClassLoader());
-
-			MethodKey methodKey = new MethodKey(
-				"com.liferay.portal.license.util.LicenseUtil", "registerOrder",
-				String.class, String.class, int.class);
-
-			PortalClassInvoker.invoke(
-				methodKey, orderUuid, productEntryName, 0);
-		}
-		finally {
-			currentThread.setContextClassLoader(classLoader);
-		}
+		LicenseUtil.registerOrder(orderUuid, productEntryName, 0);
 	}
 
 	private static final String _LICENSE_SERVER_ID_FILE_NAME =
