+++ modules/apps/marketplace/marketplace-web/docroot/WEB-INF/src/com/liferay/marketplace/web/util/PortletPropsValues.java
@@ -12,16 +12,22 @@
  * details.
  */
 
-package com.liferay.marketplace.util;
+package com.liferay.marketplace.web.util;
+
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.util.portlet.PortletProps;
 
 /**
  * @author Ryan Park
  */
-public interface PortletPropsKeys {
+public class PortletPropsValues {
 
-	public static final String MARKETPLACE_STORE_ENABLED =
-		"marketplace.store.enabled";
+	public static final boolean MARKETPLACE_STORE_ENABLED =
+		GetterUtil.getBoolean(
+			PortletProps.get(PortletPropsKeys.MARKETPLACE_STORE_ENABLED), true);
 
-	public static final String MARKETPLACE_URL = "marketplace.url";
+	public static final String MARKETPLACE_URL = GetterUtil.getString(
+		PortletProps.get(PortletPropsKeys.MARKETPLACE_URL),
+		"https://mp.liferay.com");
 
 }
